requirements.txt
Django>=4.2
django-vite>=2
mysqlclient>=2
Pillow>=9
python-dotenv